pyproject.toml
[tool.poetry]
name = "sc2bot"
version = "0.1.0"
description = ""
authors = ["Leon Leibmann <leon.leibmann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lxml = "^4.9.3"
cachetools = "^5.3.2"
discord-py = "^2.3.2"
pytz = "^2023.3.post1"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
requests = "^2.32.3"
anyascii = "^0.3.2"
python-levenshtein = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
